requirements.txt
hydra-core~=1.3.2
nmslib~=2.1.1
numpy~=1.26.4
omegaconf~=2.3.0
pandas~=2.2.1
pytorch-lightning~=2.2.1
pytorch-metric-learning~=2.4.1
ranx~=0.3.19
retriv~=0.2.3
scipy~=1.12.0
seaborn~=0.13.2
tensorboard~=2.16.2
torch~=2.2.1
torchmetrics~=1.3.1
tqdm~=4.66.2
transformers~=4.38.2
wandb~=0.16.4